pyproject.toml
[project]
name = "cotrader"
version = "0.1.0"
description = ""
authors = [{ name = "jakmobius", email = "jakmobius@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "dateparser (>=1.2.1,<2.0.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "python-binance (>=1.0.28,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "dvc (>=3.59.1,<4.0.0)",
    "lightning (>=2.5.1.post0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "jsonargparse[signatures] (>=4.39.0,<5.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "standard-imghdr (>=3.10.0,<4.0.0)",
    "typing-extensions (>=4.14.0,<5.0.0)",
    "sanic (>=25.3.0,<26.0.0)",
    "sanic-ext (>=24.12.0,<25.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "mlflow (>=2.22.1,<3.0.0)",
    "lightning-bolts (>=0.7.0,<0.8.0)",
]

[tool.poetry]
packages = [{ include = "cotrader" }]

[tool.isort]
profile = 'black'

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
